blog 100%

+++ src/main/java/com/zhong/web/admin/BlogController.java
@@ -107,8 +107,13 @@ public class BlogController {
 
     @GetMapping("/blogs/{id}/delete")
     public String delete(@PathVariable Long id,RedirectAttributes attributes){
-        blogService.deleteBlog(id);
-        attributes.addFlashAttribute("message","删除成功");
+        Blog blog = blogService.getBlog(id);
+        if (blog.isCommentabled()){
+            attributes.addFlashAttribute("message","评论存在，无法删除");
+        } else {
+            blogService.deleteBlog(id);
+            attributes.addFlashAttribute("message","删除成功");
+        }
         return REDIRECT_LIST;
     }
